fix: FP64 dtype propagation — fix silent FP32 truncation in 2 paths
Code review found 2 dtype mismatch issues from the FP64 integral change:

1. molecular.py:1699 — general-case matrix_elements() created FP32 matrix,
   silently truncating FP64 diagonal/off-diagonal elements on assignment.
   Fixed: use dtype=self.h1e.dtype (FP64) matching the fast path.

2. physics_guided_training.py:900 — off_diag buffer was FP32, truncating
   FP64 weighted elements during scatter_add_. Fixed: match weighted.dtype.

+++ src/flows/physics_guided_training.py
@@ -897,7 +897,7 @@ class PhysicsGuidedFlowTrainer:
             weighted = all_elements * ratios
 
             # Step 9: Accumulate off-diagonal contributions using scatter_add
-            off_diag = torch.zeros(n_configs, device=self.device)
+            off_diag = torch.zeros(n_configs, dtype=weighted.dtype, device=self.device)
             off_diag.scatter_add_(0, all_orig_indices, weighted)
 
             # Step 10: Total local energy = diagonal + off-diagonal
